Change type of similiaryThreshold from int to Float.

+++ ontology-common/src/main/java/uk/ac/ebi/quickgo/ontology/common/coterm/CoTermRepository.java
@@ -25,5 +25,5 @@ public interface CoTermRepository extends Repository<List<CoTerm>, String> {
      * @return a list of objects, each one of which represent a GO Term that is used to annotate the same gene
      * product as the id. Each object holds statistics related to that co-occurrence.
      */
-    List<CoTerm> findCoTerms(String id, CoTermSource source, int limit, int similarityThreshold);
+    List<CoTerm> findCoTerms(String id, CoTermSource source, int limit, Float similarityThreshold);
 }
